git action plugin now uses component version
Previously, the git action plugin derived the commit ID from the
component name, assuming the component name was of the following
formats:

<commit_id>
<component_typ>=<commit_id>

This should not be the case after the components were redesigned and the
"version" field was added.

With this CR, the git action plugin retrieves the commit ID directly
from the component version field.

Change-Id: I5e14f7078477dfe0133b7407b9e69213a69d3c65