Avoid reencoding twice the ansible tasks in the backlog
Until we have a self._job_id set by the dci_job task, all ansibles tasks are encoded and stored in self._backlog.
When we process the backlog after the first dci_job, all tasks stored in the backlog are reencoded.
Python raise an exception because we can't encode a binary string. The backlog is never uploaded.

This patch encode task content only before we create the task on the server.

Change-Id: Ibf9593ab3f1eb9641379cd1b3701f824ee17f5c5