Avoid reencoding twice the ansible tasks in the backlog
Until we have a self._job_id set by the dci_job task, all ansibles tasks are stored in self._backlog but the content in binary.
When we process the backlog after the first dci_job, all the content stored in the backlog are reencoded.
Python raise an exception and the backlog is never uploaded.

Change-Id: Ibf9593ab3f1eb9641379cd1b3701f824ee17f5c5
TestBos2: virt control-plane